+++ projects/mainio-forms/src/lib/interfaces/i-validation-message.ts
@@ -1,0 +1,5 @@
+export interface IValidationMessage {
+  error: string;
+  message: string;
+  paramValues?: string[];
+}