+++ projects/mainio-forms/src/lib/shared-components/dynamic-form-question/dynamic-form-question.component.ts
@@ -4,26 +4,31 @@ import {
   OnInit,
   OnChanges,
   SimpleChanges,
-  ChangeDetectionStrategy
+  ChangeDetectionStrategy,
+  OnDestroy
 } from "@angular/core";
 import { FormGroup, FormControl, AbstractControl } from "@angular/forms";
 
 import { QuestionBase } from "../../models/question-base";
 import { DropdownSearchQuestion } from "../../models/drop-down-search";
-import { Observable, Subject } from "rxjs";
+import { Observable, Subject, of, Subscription } from "rxjs";
 import { startWith, map } from "rxjs/operators";
 import { DropdownQuestion } from "../../models/dropdown-question";
 import { ControlType } from "../../models/control-type.enum";
 import { RepeatInput } from "../../models/repeat-input";
 import { ILoadedValues } from "../../interfaces";
 import { QuestionCreatorService } from "../../services/question-creator.service";
+import { ValidationMessagesService } from "../../services/validation-messages.service";
+import { LibraryLoggerService } from "../../services/library-logger.service";
 
 @Component({
   selector: "mainio-form-question",
   styleUrls: ["./dynamic-form-question.component.css"],
   templateUrl: "./dynamic-form-question.component.html"
 })
-export class DynamicFormQuestionComponent implements OnChanges {
+export class DynamicFormQuestionComponent implements OnChanges, OnDestroy {
+  @Input()
+  formId: string;
   @Input()
   compactStyle: boolean;
   @Input()
@@ -32,11 +37,16 @@ export class DynamicFormQuestionComponent implements OnChanges {
   form: FormGroup;
   @Input()
   values: ILoadedValues;
-
-  controller: AbstractControl;
+  errors: string[];
+  errorsSubscription: Subscription;
+  controller: FormControl;
   public shallowQuestion: QuestionBase<any>;
 
-  constructor(private _creator: QuestionCreatorService) {}
+  constructor(
+    private _creator: QuestionCreatorService,
+    private _validationMessage: ValidationMessagesService,
+    private _logger: LibraryLoggerService
+  ) {}
 
   get controlTypeString() {
     switch (this.question.controlType.toString()) {
@@ -78,11 +88,15 @@ export class DynamicFormQuestionComponent implements OnChanges {
     });
   }
 
+  ngOnDestroy() {
+    if (this.errorsSubscription) this.errorsSubscription.unsubscribe();
+  }
   ngOnChanges(changes: SimpleChanges) {
     this.shallowQuestion = this._creator.createQuestionFromControlType(
       this.question.controlType,
       this.question
     );
+
     if (this.values) {
       this.shallowQuestion.setValue(
         this.values.values[this.question.key]
@@ -90,7 +104,31 @@ export class DynamicFormQuestionComponent implements OnChanges {
           : this.question.value
       );
     }
-    this.controller = this.form.controls[this.question.key];
+    this.controller = this.form.controls[this.question.key] as FormControl;
+    if (!this.controller) {
+      this._logger.error(
+        "Form control missing for question container. Tried to locate control from FormGroup with key " +
+        !this.question
+          ? "undefined question"
+          : !this.question.key
+            ? "undefined"
+            : this.question.key
+      );
+      return;
+    }
+    this.errors = this._validationMessage.getValidationMessages(
+      this.controller as FormControl,
+      this.question.key,
+      this.formId
+    );
+    this.errorsSubscription = this.controller.statusChanges.subscribe(x => {
+      this.errors = this._validationMessage.getValidationMessages(
+        this.controller as FormControl,
+        this.question.key,
+        this.formId
+      );
+    });
+
     this.controller.setValue(this.shallowQuestion.value, {
       emitEvent: false,
       onlySelf: true