+++ projects/mainio-forms/src/lib/shared-components/form-fields/single-line-input/single-line-input.component.ts
@@ -22,6 +22,8 @@ export class SingleLineInputComponent extends FormFieldBaseComponent {
   question: InputQuestion;
   @Input()
   controller: AbstractControl;
+  @Input()
+  errors: string[];
   minLength: number = 0;
   maxLength: number = 0;
   inputHint: string = "";